fix(app): hash the whole input instead of its first 64 KiB
The Rust declaration of hash_sha256 typed its length as u16 while the C
prototype takes a uint32_t, and crypto_sha256_one_shot narrowed it again on
the way to the hasher. Any input longer than 65535 bytes was silently
truncated to its remainder modulo 64 KiB.

Every target but Nano S buffers 85 KiB, so a SIP-018 structured message can
reach that size, and its hash is what gets signed. The device would return a
signature that is valid over a prefix of the message it was handed, while the
review showed a digest that matched only that prefix. Widening the length to
uint32_t end to end makes the signature cover what was actually sent.

+++ app/src/crypto_helper.h
@@ -32,9 +32,9 @@ extern "C" {
 #define HRP            "sov"
 
 zxerr_t crypto_sha256_init();
-zxerr_t crypto_sha256_update(const uint8_t *input, uint16_t inputLen);
+zxerr_t crypto_sha256_update(const uint8_t *input, uint32_t inputLen);
 zxerr_t crypto_sha256_final(uint8_t *output);
-zxerr_t crypto_sha256_one_shot(uint8_t *output, uint16_t outputLen, const uint8_t *input, uint16_t inputLen);
+zxerr_t crypto_sha256_one_shot(uint8_t *output, uint16_t outputLen, const uint8_t *input, uint32_t inputLen);
 zxerr_t crypto_computeAddress(uint8_t *address, uint16_t addressLen, const uint8_t *pubkey);
 
 #ifdef __cplusplus
